+++ CHANGELOG.md
@@ -121,3 +121,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [3.0.19] - 2017-01-24 13:00:00
 ### Added
 - Option to fetch variant attributes by frontend_label if store_label does not exist
+
+## [3.1.0] - 2017-01-24 13:00:00
+### Added
+- Bundle products support
+- Alternative way to get product url
+- Send emails template